requirements.txt
typer>=0.9.0
pyyaml>=6.0
pytest>=7.0
